update istio manifests
