--- internal/assets/manifests/base/templates/clusterrolebinding.yaml
@@ -1,32 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: {{ include "name-with-namespaced-revision" ( dict "name" "istio-reader" "context" $) }}
-  labels:
-    app: istio-reader
-    release: {{ .Release.Name }}
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: {{ include "name-with-namespaced-revision" ( dict "name" "istio-reader" "context" $) }}
-subjects:
-  - kind: ServiceAccount
-    name: {{ include "name-with-revision" ( dict "name" "istio-reader-service-account" "context" $) }}
-    namespace: {{ .Release.Namespace }}
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: {{ include "name-with-namespaced-revision" ( dict "name" "istiod" "context" $) }}
-  labels:
-    app: istiod
-    release: {{ .Release.Name }}
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: {{ include "name-with-namespaced-revision" ( dict "name" "istiod" "context" $) }}
-subjects:
-  - kind: ServiceAccount
-    name: {{ include "name-with-revision" ( dict "name" "istiod-service-account" "context" $) }}
-    namespace: {{ .Values.global.istioNamespace }}
----
