+++ internal/assets/manifests/base/templates/endpoints.yaml
@@ -1,5 +1,5 @@
 {{- if .Values.global.remotePilotAddress }}
-  {{- if .Values.pilot.enabled }}
+  {{- if not .Values.global.externalIstiod }}
 apiVersion: v1
 kind: Endpoints
 metadata:
