+++ internal/assets/manifests/base/templates/services.yaml
@@ -1,5 +1,5 @@
 {{- if .Values.global.remotePilotAddress }}
-  {{- if .Values.pilot.enabled }}
+  {{- if not .Values.global.externalIstiod }}
 # when istiod is enabled in remote cluster, we can't use istiod service name
 apiVersion: v1
 kind: Service
