+++ internal/assets/manifests/istio-cni/templates/daemonset.yaml
@@ -43,7 +43,6 @@ spec:
     spec:
       nodeSelector:
         kubernetes.io/os: linux
-      hostNetwork: true
       tolerations:
         # Make sure istio-cni-node gets scheduled on all nodes.
         - effect: NoSchedule
@@ -97,45 +96,34 @@ spec:
             # Deploy as a standalone CNI plugin or as chained?
             - name: CHAINED_CNI_PLUGIN
               value: "{{ .Values.cni.chained }}"
+            - name: REPAIR_ENABLED
+              value: "{{ .Values.cni.repair.enabled }}"
+            - name: REPAIR_NODE_NAME
+              valueFrom:
+                fieldRef:
+                  fieldPath: spec.nodeName
+            - name: REPAIR_LABEL_PODS
+              value: "{{.Values.cni.repair.labelPods}}"
+            # Set to true to enable pod deletion
+            - name: REPAIR_DELETE_PODS
+              value: "{{.Values.cni.repair.deletePods}}"
+            - name: REPAIR_RUN_AS_DAEMON
+              value: "true"
+            - name: REPAIR_SIDECAR_ANNOTATION
+              value: "sidecar.istio.io/status"
+            - name: REPAIR_INIT_CONTAINER_NAME
+              value: "{{ .Values.cni.repair.initContainerName }}"
+            - name: REPAIR_BROKEN_POD_LABEL_KEY
+              value: "{{.Values.cni.repair.brokenPodLabelKey}}"
+            - name: REPAIR_BROKEN_POD_LABEL_VALUE
+              value: "{{.Values.cni.repair.brokenPodLabelValue}}"
           volumeMounts:
             - mountPath: /host/opt/cni/bin
               name: cni-bin-dir
             - mountPath: /host/etc/cni/net.d
               name: cni-net-dir
-{{- if .Values.cni.repair.enabled }}
-        - name: repair-cni
-{{- if contains "/" .Values.cni.image }}
-          image: "{{ .Values.cni.image }}"
-{{- else }}
-          image: "{{ .Values.cni.hub | default .Values.global.hub }}/{{ .Values.cni.image | default "install-cni" }}:{{ .Values.cni.tag | default .Values.global.tag }}"
-{{- end }}
-{{- if or .Values.cni.pullPolicy .Values.global.imagePullPolicy }}
-          imagePullPolicy: {{ .Values.cni.pullPolicy | default .Values.global.imagePullPolicy }}
-{{- end }}
-
-          command: ["/opt/local/bin/istio-cni-repair"]
-          env:
-          - name: "REPAIR_NODE-NAME"
-            valueFrom:
-              fieldRef:
-                fieldPath: spec.nodeName
-          - name: "REPAIR_LABEL-PODS"
-            value: "{{.Values.cni.repair.labelPods}}"
-          # Set to true to enable pod deletion
-          - name: "REPAIR_DELETE-PODS"
-            value: "{{.Values.cni.repair.deletePods}}"
-          - name: "REPAIR_RUN-AS-DAEMON"
-            value: "true"
-          - name: "REPAIR_SIDECAR-ANNOTATION"
-            value: "sidecar.istio.io/status"
-          - name: "REPAIR_INIT-CONTAINER-NAME"
-            value: "{{ .Values.cni.repair.initContainerName }}"
-          - name: "REPAIR_BROKEN-POD-LABEL-KEY"
-            value: "{{.Values.cni.repair.brokenPodLabelKey}}"
-          - name: "REPAIR_BROKEN-POD-LABEL-VALUE"
-            value: "{{.Values.cni.repair.brokenPodLabelValue}}"
-{{- end }}
-
+            - mountPath: /var/run/istio-cni
+              name: cni-log-dir
 {{- if .Values.cni.taint.enabled }}
         - name: taint-controller
 {{- if contains "/" .Values.cni.image }}
@@ -163,3 +151,7 @@ spec:
         - name: cni-net-dir
           hostPath:
             path: {{ default "/etc/cni/net.d" .Values.cni.cniConfDir }}
+        # Used for UDS log
+        - name: cni-log-dir
+          hostPath:
+            path: /var/run/istio-cni
