+++ internal/assets/manifests/istio-control/istio-discovery/templates/clusterrolebinding.yaml
@@ -1,16 +1,15 @@
 apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
+kind: ClusterRoleBinding
 metadata:
   name: {{ include "name-with-namespaced-revision" ( dict "name" "istiod" "context" $) }}
-  namespace: {{ .Release.Namespace }}
   labels:
     app: istiod
     release: {{ .Release.Name }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
-  kind: Role
+  kind: ClusterRole
   name: {{ include "name-with-namespaced-revision" ( dict "name" "istiod" "context" $) }}
 subjects:
   - kind: ServiceAccount
-    name: {{ include "name-with-revision" ( dict "name" "istiod-service-account" "context" $) }}
-    namespace: {{ .Release.Namespace }}
+    name: {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
+    namespace: {{ .Values.global.istioNamespace }}
