+++ internal/assets/manifests/istio-control/istio-discovery/templates/configmap-jwks.yaml
@@ -7,8 +7,6 @@ metadata:
   labels:
     release: {{ .Release.Name }}
     istio.io/rev: {{ include "namespaced-revision" . }}
-    install.operator.istio.io/owning-resource: {{ .Values.ownerName | default "unknown" }}
-    operator.istio.io/component: "Pilot"
 data:
   extra.pem: {{ .Values.pilot.jwksResolverExtraRootCA | quote }}
 {{- end }}
