--- internal/assets/manifests/istio-control/istio-discovery/templates/meshexpansion.yaml
@@ -1,70 +1,0 @@
-apiVersion: networking.istio.io/v1alpha3
-kind: Gateway
-metadata:
-  name: cross-network-gateway
-spec:
-  selector:
-    istio: eastwestgateway
-  servers:
-    - port:
-        number: 15443
-        name: tls
-        protocol: TLS
-      tls:
-        mode: AUTO_PASSTHROUGH
-      hosts:
-        - "*.local"
----
-apiVersion: networking.istio.io/v1alpha3
-kind: Gateway
-metadata:
-  name: istiod-gateway
-spec:
-  selector:
-    istio: eastwestgateway
-  servers:
-    - port:
-        name: tls-istiod
-        number: 15012
-        protocol: tls
-      tls:
-        mode: PASSTHROUGH
-      hosts:
-        - "*"
-    - port:
-        name: tls-istiodwebhook
-        number: 15017
-        protocol: tls
-      tls:
-        mode: PASSTHROUGH
-      hosts:
-        - "*"
----
-apiVersion: networking.istio.io/v1alpha3
-kind: VirtualService
-metadata:
-  name: istiod-vs
-spec:
-  hosts:
-  - "*"
-  gateways:
-  - istiod-gateway
-  tls:
-  - match:
-    - port: 15012
-      sniHosts:
-      - "*"
-    route:
-    - destination:
-        host: istiod.istio-system.svc.cluster.local
-        port:
-          number: 15012
-  - match:
-    - port: 15017
-      sniHosts:
-      - "*"
-    route:
-    - destination:
-        host: istiod.istio-system.svc.cluster.local
-        port:
-          number: 443
