+++ internal/assets/manifests/istio-control/istio-discovery/templates/mutatingwebhook.yaml
@@ -5,12 +5,16 @@ a unique prefix to each. */}}
 - name: {{.Prefix}}sidecar-injector.istio.io
   clientConfig:
     {{- if .Values.istiodRemote.injectionURL }}
-    url: {{ .Values.istiodRemote.injectionURL }}
+    url: "{{ .Values.istiodRemote.injectionURL }}"
     {{- else }}
     service:
       name: {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
       namespace: {{ .Release.Namespace }}
+      {{- if .Values.istiodRemote.injectionURL }}
       path: "/inject"
+      {{- else }}
+      path: "{{ .Values.istiodRemote.injectionPath }}"
+      {{- end }}
       port: 443
     {{- end }}
     caBundle: ""
@@ -36,65 +40,10 @@ metadata:
   {{- end }}
   labels:
     istio.io/rev: {{ include "namespaced-revision" . }}
-    install.operator.istio.io/owning-resource: {{ .Values.ownerName | default "unknown" }}
-    operator.istio.io/component: "Pilot"
     app: sidecar-injector
     release: {{ .Release.Name }}
 webhooks:
-{{- if .Values.sidecarInjectorWebhook.useLegacySelectors}}
-{{- /* Setup the "legacy" selectors. These are for backwards compatibility, will be removed in the future. */}}
-{{- include "core" . }}
-  namespaceSelector:
-  {{- if .Values.sidecarInjectorWebhook.enableNamespacesByDefault }}
-    matchExpressions:
-    - key: name
-      operator: NotIn
-      values:
-      - {{ .Release.Namespace }}
-    - key: istio-injection
-      operator: NotIn
-      values:
-      - disabled
-    - key: istio-env
-      operator: DoesNotExist
-    - key: istio.io/rev
-      operator: DoesNotExist
-  {{- else if .Values.revision }}
-    matchExpressions:
-    - key: istio-injection
-      operator: DoesNotExist
-    - key: istio.io/rev
-      operator: In
-      values:
-      - {{ include "namespaced-revision" . }}
-  {{- else }}
-    matchLabels:
-      istio-injection: enabled
-  {{- end }}
-  {{- if .Values.sidecarInjectorWebhook.objectSelector.enabled }}
-  objectSelector:
-    {{- if .Values.sidecarInjectorWebhook.objectSelector.autoInject }}
-    matchExpressions:
-    - key: "sidecar.istio.io/inject"
-      operator: NotIn
-      values:
-      - "false"
-    {{- else if .Values.revision }}
-    matchExpressions:
-    - key: "sidecar.istio.io/inject"
-      operator: DoesNotExist
-    - key: istio.io/rev
-      operator: In
-      values:
-      - {{ include "namespaced-revision" . }}
-    {{- else }}
-    matchLabels:
-      "sidecar.istio.io/inject": "true"
-    {{- end }}
-  {{- end }}
-{{- else }}
-
-  {{- /* Set up the selectors. First section is for revision, rest is for "default" revision */}}
+{{- /* Set up the selectors. First section is for revision, rest is for "default" revision */}}
 
 {{- /* Case 1: namespace selector matches, and object doesn't disable */}}
 {{- /* Note: if both revision and legacy selector, we give precedence to the legacy one */}}
@@ -104,11 +53,7 @@ webhooks:
     - key: istio.io/rev
       operator: In
       values:
-      {{- if (eq .Values.revision "") }}
-      - "default"
-      {{- else }}
-      - {{ include "namespaced-revision" . }}
-      {{- end }}
+      - "{{ include "namespaced-revision" . }}"
     - key: istio-injection
       operator: DoesNotExist
   objectSelector:
@@ -135,11 +80,7 @@ webhooks:
     - key: istio.io/rev
       operator: In
       values:
-      {{- if (eq .Values.revision "") }}
-      - "default"
-      {{- else }}
       - {{ include "namespaced-revision" . }}
-      {{- end }}
 
 
 {{- /* Webhooks for default revision */}}
@@ -194,6 +135,5 @@ webhooks:
       operator: DoesNotExist
 {{- end }}
 
-{{- end }}
 {{- end }}
 {{- end }}
