+++ internal/assets/manifests/istio-control/istio-discovery/templates/role.yaml
@@ -1,7 +1,7 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: Role
 metadata:
-  name: {{ include "name-with-namespaced-revision" ( dict "name" "istiod" "context" $) }}
+  name:  {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
   namespace: {{ .Release.Namespace }}
   labels:
     app: istiod
