+++ internal/assets/manifests/istio-control/istio-discovery/templates/rolebinding.yaml
@@ -1,0 +1,16 @@
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  name:  {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app: istiod
+    release: {{ .Release.Name }}
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: Role
+  name:  {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
+subjects:
+  - kind: ServiceAccount
+    name: {{ include "name-with-revision" ( dict "name" "istiod" "context" $) }}
+    namespace: {{ .Release.Namespace }}
