+++ internal/assets/manifests/istio-control/istio-discovery/values.yaml
@@ -33,17 +33,7 @@ pilot:
   enableProtocolSniffingForInbound: true
 
   nodeSelector: {}
-  affinity: {}
-  tolerations: {}
   podAnnotations: {}
-
-  securityContext:
-    runAsUser: 1337
-    runAsGroup: 1337
-    runAsNonRoot: true
-    capabilities:
-      drop:
-        - ALL
 
   # You can use jwksResolverExtraRootCA to provide a root certificate
   # in PEM format. This will then be trusted by pilot when resolving
@@ -75,10 +65,6 @@ pilot:
 
 
 sidecarInjectorWebhook:
-  # If enabled, the legacy webhook selection logic will be used. This relies on filtering of webhook
-  # requests in Istiod, rather than at the webhook selection level.
-  # This is option is intended for migration purposes only and will be removed in Istio 1.10.
-  useLegacySelectors: false
   # You can use the field called alwaysInjectSelector and neverInjectSelector which will always inject the sidecar or
   # always skip the injection on pods that match that label selector, regardless of the global policy.
   # See https://istio.io/docs/setup/kubernetes/additional-setup/sidecar-injection/#more-control-adding-exceptions
@@ -140,9 +126,15 @@ sidecarInjectorWebhook:
   # defaultTemplates: ["sidecar", "hello"]
   defaultTemplates: []
 istiodRemote:
-  # Sidecar injector mutating webhook configuration url
+  # Sidecar injector mutating webhook configuration clientConfig.url value.
   # For example: https://$remotePilotAddress:15017/inject
+  # The host should not refer to a service running in the cluster; use a service reference by specifying
+  # the clientConfig.service field instead.
   injectionURL: ""
+
+  # Sidecar injector mutating webhook configuration path value for the clientConfig.service field.
+  # Override to pass env variables, for example: /inject/cluster/remote/net/network2
+  injectionPath: "/inject"
 telemetry:
   enabled: true
   v2:
@@ -246,9 +238,9 @@ global:
   # Default hub for Istio images.
   # Releases are published to docker hub under 'istio' project.
   # Dev builds from prow are on gcr.io
-  hub: gcr.io/istio-testing
+  hub: istio
   # Default tag for Istio images.
-  tag: latest
+  tag: "1.11.0"
 
   # Specify image pull policy if default behavior isn't desired.
   # Default behavior: latest images will be Always else IfNotPresent.
@@ -397,8 +389,13 @@ global:
   # If not set explicitly, default to the Istio discovery address.
   caAddress: ""
 
-  # External istiod controls all remote clusters: disabled by default
+  # Configure a remote cluster data plane controlled by an external istiod.
+  # When set to true, istiod is not deployed locally and only a subset of the other
+  # discovery charts are enabled.
   externalIstiod: false
+
+  #  Configure a remote cluster as the config cluster for an external istiod.
+  configCluster: false
 
   # Configure the policy for validating JWT.
   # Currently, two options are supported: "third-party-jwt" and "first-party-jwt".
@@ -520,6 +517,9 @@ global:
 
   # Use the Mesh Control Protocol (MCP) for configuring Istiod. Requires an MCP source.
   useMCP: false
+
+  # Determines whether this istiod performs resource validation.
+  configValidation: true
 
 base:
   # For istioctl usage to disable istio config crds in base
